+++ rest/rest.go
@@ -133,6 +133,7 @@ func (r *RESTD) Run(u uhppote.IUHPPOTE, devices []uhppote.Device) {
 			handler{regexp.MustCompile("^/uhppote/device/[0-9]+/door/[1-4]/delay$"), http.MethodPut, device.SetDoorDelay},
 			handler{regexp.MustCompile("^/uhppote/device/[0-9]+/door/[1-4]/control$"), http.MethodPut, device.SetDoorControl},
 			handler{regexp.MustCompile("^/uhppote/device/[0-9]+/door/interlock$"), http.MethodPut, device.SetInterlock},
+			handler{regexp.MustCompile("^/uhppote/device/[0-9]+/door/keypads$"), http.MethodPut, device.ActivateKeypads},
 
 			handler{regexp.MustCompile("^/uhppote/device/[0-9]+/cards$"), http.MethodGet, device.GetCards},
 			handler{regexp.MustCompile("^/uhppote/device/[0-9]+/cards$"), http.MethodDelete, device.DeleteCards},